Sensors y actuator controllers

+++ app/Http/Controllers/UsersController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
+use Illuminate\Support\Facades\Hash;
+
 
 class UsersController extends Controller
 {
@@ -18,20 +20,36 @@ class UsersController extends Controller
 
 //crear un usuario
     public function store(Request $request){
-        return User::create($request->all());
-    }
-
-    //actualizar un usuario
-    public function update(Request $request, $id){
-        $user = User::find($id);
-        $user->update($request->all());
+        $this->validate($request, [
+            'username' => 'required|unique:users',
+            'password' => 'required',
+            'rol' => 'required',
+        ]);
+        $user = new User;
+        $user->fill($request->all());
+        $user->password = Hash::make($request->password);
         $user->save();
         return $user;
     }
 
-     //actualizar un usuario
+    //actualizar un usuario
+    public function update(Request $request, $id){
+        $this->validate($request, [
+            'username' => 'filled|unique:users',
+        ]);
+        $user = User::find($id);
+        if(!$user)return response('', 404);
+        $user->update($request->all());
+        if($request->password)
+            $user->password = Hash::make($request->password);
+        $user->save();
+        return $user;
+    }
+
+     //borrar un usuario
      public function destroy($id){
         $user = User::find($id);
+        if(!$user)return response('', 404);
         $user->delete();
         return $user;
     }
